Conflicts:
	usci/db_oracle_scripts/eav_be_entities_list_partitioning.sql

+++ usci/modules/eav/modules/persistance/src/main/java/kz/bsbnb/usci/eav/stats/QueryEntry.java
@@ -3,8 +3,8 @@ package kz.bsbnb.usci.eav.stats;
 import java.io.Serializable;
 
 public class QueryEntry implements Serializable {
-    public double maxTime;
-    public double minTime;
-    public double totalTime;
+    public long maxTime;
+    public long minTime;
+    public long totalTime;
     public long count;
 }
